% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_mae}
\alias{create_mae}
\title{Create multi-assay experiment object.}
\usage{
create_mae(study_name = "TCGA")
}
\arguments{
\item{study_name}{study identifier.}
}
\value{
A MulitAssayExperiment object containing clinical and gene expression data 
along with any other available data
}
\description{
Create multi-assay experiment object.
}
