% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedPCaData-package.R
\docType{package}
\name{curatedPCaData}
\alias{curatedPCaData}
\alias{curatedPCaData-package}
\title{curatedPCaData}
\description{
Description
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Syksy/curatedPCaData}
  \item Report bugs at \url{https://github.com/Syksy/curatedPCaData/issues}
}

}
\author{
\strong{Maintainer}: Teemu Daniel Laajala \email{teelaa@utu.fi} (\href{https://orcid.org/0000-0002-7016-7354}{ORCID})

Other contributors:
\itemize{
  \item Jordan Creed \email{jordan.h.creed@moffitt.org} [contributor]
  \item Christelle Colin Leitzinger \email{christelle.colinleitzinger@moffitt.org} [contributor]
  \item Varsha Sreekanth \email{varsha.sreekanth@cuanschutz.edu} [contributor]
  \item Federico Calboli \email{federico.calboli@utu.fi} [contributor]
  \item Kalaimathy Singaravelu \email{kalsin@utu.fi} [contributor]
  \item Michael Orman \email{michael.orman@cuanschutz.edu} [contributor]
  \item Alex Soupir \email{Alex.Soupir@moffitt.org} [contributor]
}

}
\keyword{internal}
