% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{curatedPCaDatasets_chandran}
\alias{curatedPCaDatasets_chandran}
\title{Chandran et al., Yu et al. MAE-object}
\format{
An MAE object spanning 171 men with prostate cancer
\describe{
    \item{gex.rma}{matrix with 9007 rows and 171 columns, for the gene 
        expression data}
    \item{cibersort}{matrix with 22 rows and 171 columns, of cibersort based 
        deconvolution data}
    \item{xcell}{matrix with 39 rows and 171 columns, of xcell based 
        deconvolution data}
    \item{epic}{matrix with 8 rows and 171 columns, of epic based 
        deconvolution data}
    \item{quantiseq}{matrix with 11 rows and 171 columns, of quantiseq based 
        deconvolution data}
    \item{estimate}{data.frame with 4 rows and 171 columns, of cell types 
        based on ESTIMATE method}
    \item{scores}{matrix with 4 rows and 171 columns, of risk scores and AR 
        scores}
    \item{mcp}{matrix with 11 rows and 171 columns, of mcp-counter based 
        deconvolution data}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE6919}
}
\value{
A MultiAssayExperiment corresponding to the study and its available 
omics.
}
\description{
MultiAssayExperiment object containing gene expression (gex) and immune cell 
estimates for Chandran et al., Yu et al.
}
\examples{
mae_chandran <- getPCa('chandran')
}
\references{
(\href{https://pubmed.ncbi.nlm.nih.gov/17430594/}{PubMed}) 
(\href{https://pubmed.ncbi.nlm.nih.gov/15254046/}{PubMed})
}
\keyword{datasets}
