% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{curatedPCaDatasets_kunderfranco}
\alias{curatedPCaDatasets_kunderfranco}
\title{Kunderfranco et al. MAE-object}
\format{
An MAE object spanning 67 samples of normal prostate samples and 
prostate cancer samples
\describe{
    \item{gex.logr}{matrix with 16546 rows and 67 columns, of gene expression
        data}
    \item{cibersort}{matrix with 22 rows and 67 columns, of cibersort based 
        deconvolution data}
    \item{xcell}{matrix with 39 rows and 67 columns, of xcell based 
        deconvolution data}
    \item{epic}{matrix with 8 rows and 67 columns, of epic based 
        deconvolution data}
    \item{quantiseq}{matrix with 11 rows and 67 columns, of quantiseq based 
        deconvolution data}
    \item{estimate}{data.frame with 4 rows and 67 columns, of cell types 
        based on ESTIMATE method}
    \item{scores}{matrix with 2 rows and 67 columns, of risk scores and AR 
        scores}
    \item{mcp}{matrix with 11 rows and 67 columns, of mcp-counter based 
        deconvolution data}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE14206}
}
\value{
A MultiAssayExperiment corresponding to the study and its available 
omics.
}
\description{
MultiAssayExperiment object containing gene expression (gex) and immune cell 
estimates from Kunderfranco et al.
}
\details{
The data contains 14 disease free benign prostate hyperplasia 
samples and 53 prostate cancer samples.
}
\examples{
mae_kunderfranco <- getPCa('kunderfranco')
}
\references{
(\href{https://pubmed.ncbi.nlm.nih.gov/20479932/}{PubMed})
}
\keyword{datasets}
