% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_osf_data.R
\name{format_osf_data}
\alias{format_osf_data}
\title{Does pre-formatting of OSF dataset so that the genes are in HGNC format and sample names are TCGA IDs}
\usage{
format_osf_data(osf_data)
}
\arguments{
\item{osf_data}{}
}
\description{
Does pre-formatting of OSF dataset so that the genes are in HGNC format and sample names are TCGA IDs
}
