% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_cbioportal}
\alias{generate_cbioportal}
\title{Download generic 'omics data from cBioPortal using dataset specific query}
\usage{
generate_cbioportal(
  genes = sort(curatedPCaData:::tcga_gene_names$hgnc),
  geneticProfiles = "prad_tcga_pub_rna_seq_v2_mrna",
  caseList = "prad_tcga_pub_sequenced",
  delay = 0.05,
  splitsize = 100,
  verb = TRUE,
  ...
)
}
\description{
Download generic 'omics data from cBioPortal using dataset specific query
}
