% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_cbioportal_oncoprint}
\alias{generate_cbioportal_oncoprint}
\title{Download mutation data from cBioPortal and format them into an oncoprint-friendly matrix}
\usage{
generate_cbioportal_oncoprint(
  study_id,
  genes = sort(unique(curatedPCaData:::curatedPCaData_genes$hgnc_symbol)),
  delay = 0.05,
  splitsize = 100,
  verb = TRUE,
  oncoprintify = TRUE
)
}
\arguments{
\item{study_id}{TODO}

\item{genes}{TODO}

\item{delay}{TODO}

\item{splitsize}{TODO}

\item{verb}{TODO}

\item{oncoprintify}{TODO}
}
\description{
Download mutation data from cBioPortal and format them into an oncoprint-friendly matrix
}
\examples{
oncop_tcga <- curatedPCaData:::generate_cbioportal_oncoprint(study_id="tcga", oncoprintify=TRUE)
oncop_taylor <- curatedPCaData:::generate_cbioportal_oncoprint(study_id="taylor", oncoprintify=TRUE)
oncop_barbieri <- curatedPCaData:::generate_cbioportal_oncoprint(study_id="barbieri", oncoprintify=TRUE)
oncop_ren <- curatedPCaData:::generate_cbioportal_oncoprint(study_id="ren", oncoprintify=TRUE)
# TODO
# oncop_abida <- curatedPCaData:::generate_cbioportal_oncoprint(study_id="abida", oncoprintify=TRUE)
}
