% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_cbioportaldata_mut}
\alias{generate_cbioportaldata_mut}
\title{Download mutation data from cBioPortal using cbioportaldata package}
\usage{
generate_cbioportaldata_mut(caselist)
}
\arguments{
\item{caselist}{TODO}
}
\description{
Download mutation data from cBioPortal using cbioportaldata package
}
\examples{
ren_mut=generate_cbioportaldata_mut("prad_eururol_2017")
barbieri_mut=generate_cbioportaldata_mut("prad_broad")
}
