% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_cgdsr_mut}
\alias{generate_cgdsr_mut}
\title{Download mutation data from cBioPortal using cgsdr package}
\usage{
generate_cgdsr_mut(
  study_id,
  genes = sort(unique(curatedPCaData:::curatedPCaData_genes$hgnc_symbol)),
  delay = 0.05,
  splitsize = 100,
  verb = TRUE
)
}
\arguments{
\item{study_id}{TODO}

\item{genes}{TODO}

\item{delay}{TODO}

\item{splitsize}{TODO}

\item{verb}{TODO}
}
\description{
Download mutation data from cBioPortal using cgsdr package
}
\examples{
ren_mut=generate_cgdsr_mut("ren",genes=curatedPCaData::curatedPCaData_genes$hgnc_symbol)
barbieri_mut=generate_cgdsr_mut("barbieri",genes=curatedPCaData::curatedPCaData_genes$hgnc_symbol)
}
