% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_gex_geo}
\alias{generate_gex_geo}
\title{Download gene expression from GEO using study specific id}
\usage{
generate_gex_geo(
  file_directory,
  geo_code = "GSE25136",
  cleanup = TRUE,
  collapse_probes = function(z) {     apply(z, MARGIN = 2, FUN = stats::median) },
  ...
)
}
\description{
Download gene expression from GEO using study specific id
}
