% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_icgc}
\alias{generate_icgc}
\title{Download and generate omics from the ICGC}
\usage{
generate_icgc(
  icgc_id = "PRAD_CA",
  set = "gex",
  file_directory,
  collapseFUN = mean,
  verb = 0
)
}
\arguments{
\item{icgc_id}{character string indicating name of ICGC dataset}

\item{file_directory}{character string indicating path for downloading the ICGC files}

\item{omic}{character string indicating which omic to download and generate for the specified icgc_id

ICGC Publication Policy for embargoes etc: http://www.icgc.org/icgc/goals-structure-policies-guidelines/e3-publication-policy
ICGC Publication guidelines: http://docs.icgc.org/portal/publication/#current-moratorium-status-for-icgc-projects . 
(e.g. 'All data shall become free of a publication moratorium when either the data is published by the ICGC member project or 
 one year after a specified quantity of data (e.g. genome dataset from 100 tumours per project) has been released via the ICGC 
 database or other public databases. In all cases data shall be free of a publication moratorium two years after its initial release.')
"PRAD-CA: No Embargo. Data available without limitations"
"PRAD-UK: No Embargo. Data available without limitations"
"PRAD-FR: No Embargo. Data available without limitations"
"PRAD-CN: Publication limitations in place until 2020-04-30" (expired) (Only simple mutations, excluded)
"EOPC-DE: No Embargo. Data available without limitations" (Need to verify biological applicability) 

NOTE: Sometimes the downloads seem to fail randomly; perhaps a fixed amount of retries ought to be allowed?}
}
\description{
Download and generate omics from the ICGC
}
