% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_xenabrowser}
\alias{generate_xenabrowser}
\title{Download GDC processed data via xenabrowser.net Santa-Cruz interface}
\usage{
generate_xenabrowser(
  id = "TCGA-PRAD",
  type = c("gex", "cna", "mut", "clinical"),
  collapse_fun = function(z) {     apply(z, MARGIN = 2, FUN = stats::median) },
  truncate = TRUE,
  digits,
  ...
)
}
\description{
Download GDC processed data via xenabrowser.net Santa-Cruz interface
}
