% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_risk.R
\name{genomic_risk}
\alias{genomic_risk}
\title{Genomic risk scores}
\usage{
genomic_risk(
  mae,
  slot = "gex",
  test = c("Prolaris", "Oncotype DX", "Decipher"),
  log = TRUE
)
}
\description{
Prolaris, Oncotype DX, Decipher genomic panels for prostate cancer risk
}
\details{
https://bjui-journals.onlinelibrary.wiley.com/doi/10.1111/bju.14452 https://bmcgenomics.biomedcentral.com/articles/10.1186/1471-2164-14-690 https://www.nature.com/articles/s41391-019-0167-9
}
