% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_risk.R
\name{genomic_score}
\alias{genomic_score}
\title{Various genomic scores}
\usage{
genomic_score(mae, slot = "gex", test = "AR", verbose = TRUE)
}
\description{
AR score by Hieronymus et al 2006 as used by TCGA 2015
}
