% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpcasummaries.R
\name{getPCaStudies}
\alias{getPCaStudies}
\title{Create a vector of unique study identifiers available in curatedPCaData}
\usage{
getPCaStudies(uniqs = TRUE)
}
\arguments{
\item{uniqs}{Should only unique instances be returned; by default TRUE}
}
\value{
A character vector with short study identifiers, e.g. 'abida', 
'baca', ...
}
\description{
This function accesses the metadata available together with the 
curatedPCaData package and creates a vector of unique study identifier names 
available in the ExperimentHub resources. This vector is useful, for example,
when applying the 'getPCa' to create MAE from the studies.
}
\details{
Uses the internal 'metadata.csv' file, which is also used to provide
the metadata
available in ExperimentHub.
}
\examples{
studyids <- getPCaStudies()

}
