% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadPCa}
\alias{loadPCa}
\title{Loop and load all MAE data objects available in curatedPCaData}
\usage{
loadPCa()
}
\value{
All data objects from curatedPCaData loaded into the environment using the 'data'-function.
}
\description{
To avoid having to use LazyLoad, the package has been set to not load all data objects upon package being loaded.
However, for many convenience reasons it'd be handy to have all the mae_*-named objects in the workspace without
having to manually iterate over all of them. This function extract the latest MAE-object list from the package and
calls 'data(mae_*)' on them.
}
\examples{
loadPCa()

}
