% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_abida}
\alias{mae_abida}
\title{Abida et al. MAE-object}
\format{

}
\source{
\url{https://www.cbioportal.org/study/summary?id=prad_su2c_2019}
}
\usage{
mae_abida
}
\description{
MultiAssayExperiment object for Abida et al.
}
\details{
the clinical data refers to a sample of 444 tumors collected in 429 patients. The tissue was collected primarily at metastatic sites rather than from the prostate.
}
\references{
Abida, W., Cyrta, J., Heller, G., et al. (2019). Genomic correlates of clinical outcome in advanced prostate cancer. Proceedings of the National Academy of Sciences of the United States of America, 116, 11428 - 11436.
(\href{https://pubmed.ncbi.nlm.nih.gov/31061129/}{PubMed})
}
\keyword{datasets}
