% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_baca}
\alias{mae_baca}
\title{Baca et al. MAE-object}
\format{
A MAE object spanning prostate cancer samples.
\describe{
  \item{cna.gistic}{matrix with 21723 rows and 109 columns, from GISTIC discretized copy number alteration calls.}
}
}
\source{
\url{...}
}
\usage{
mae_baca
}
\description{
MultiAssayExperiment object for Baca et al.
}
\keyword{datasets}
