% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_barwick}
\alias{mae_barwick}
\title{Barwick et al. MAE-object}
\format{
A MAE object spanning 146 prostate cancer samples.
\describe{
  \item{gex.logq}{matrix with 482 rows and 146 columns, for the log-quantile normalized gene expression data}
  \item{cibersort}{matrix with 22 rows and 146 columns, of cibersort based deconvolution data}
  \item{quantiseq}{matrix with 11 rows and 146 columns, of quantiseq based deconvolution data}
  \item{mcp}{matrix with 4 rows and 146 columns, of mcp-counter based deconvolution data}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE18655}
}
\usage{
mae_barwick
}
\description{
MultiAssayExperiment object containing gene expression (gex) and immune cell estimates for Barwick et al.
}
\details{
NA
}
\references{
Barwick, B. G., Abramovitz, M., Kodani, M., Moreno, C. S., Nam, R., Tang, W., Bouzyk, M., Seth, A., & Leyland-Jones, B. (2010). Prostate cancer genes associated with TMPRSS2-ERG gene fusion and prognostic of biochemical recurrence in multiple cohorts. British journal of cancer, 102(3), 570–576. https://doi.org/10.1038/sj.bjc.6605519
(\href{https://pubmed.ncbi.nlm.nih.gov/22610119/}{PubMed})
}
\keyword{datasets}
