% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_chandran}
\alias{mae_chandran}
\title{Chandran et al., Yu et al. MAE-object}
\format{
An MAE object spanning 112 men with prostate cancer
\describe{
  \item{gex}{matrix with 15681 rows and 503 columns, for the gene expression data}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE6919}
}
\usage{
mae_chandran
}
\description{
MultiAssayExperiment object for Chandran et al., Yu et al.
}
\details{
.
}
\references{
Chandran, U. R., Ma, C., Dhir, R., et al. (2007). Gene expression profiles of prostate cancer reveal involvement of multiple molecular pathways in the metastatic process. BMC cancer, 7, 64. https://doi.org/10.1186/1471-2407-7-64
(\href{https://pubmed.ncbi.nlm.nih.gov/17430594/}{PubMed})
Yu, Y. P., Landsittel, D., Jing, L., et al. (2004). Gene expression alterations in prostate cancer predicting tumor aggression and preceding development of malignancy. Journal of clinical oncology : official journal of the American Society of Clinical Oncology, 22(14), 2790–2799. https://doi.org/10.1200/JCO.2004.05.158
(\href{https://pubmed.ncbi.nlm.nih.gov/15254046/}{PubMed})
}
\keyword{datasets}
