% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_friedrich}
\alias{mae_friedrich}
\title{Friedrich et al. MAE-object}
\format{
An object of class \code{MultiAssayExperiment} of length 1.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE134160}
}
\usage{
mae_friedrich
}
\description{
MultiAssayExperiment object containing GEX from Friedrich et al.
}
\keyword{datasets}
