% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_hieronymus}
\alias{mae_hieronymus}
\title{Hieronymus et al. MAE-object}
\format{
An object of class \code{MultiAssayExperiment} of length 1.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE54691}
}
\usage{
mae_hieronymus
}
\description{
MultiAssayExperiment object containing CNA from Hieronymus et al.
}
\keyword{datasets}
