% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_icgcca}
\alias{mae_icgcca}
\title{ICGC CA MAE-object}
\format{
An object of class \code{MultiAssayExperiment} of length 1.
}
\source{
\url{https://dcc.icgc.org/projects/PRAD-CA}
}
\usage{
mae_icgcca
}
\description{
MultiAssayExperiment object containing GEX from the Canadian ICGC cohort.
}
\keyword{datasets}
