% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_kim}
\alias{mae_kim}
\title{Kim et al.}
\format{
An MAE object spanning 266 men with prostate cancer
\describe{
\item{gex.rma}{matrix with 17638 rows and 266 columns, of gene expression data}
\item{cibersort}{matrix with 22 rows and 266 columns, of cibersort based deconvolution data}
\item{xcell}{matrix with 39 rows and 266 columns, of xcell based deconvolution data}
\item{epic}{matrix with 8 rows and 266 columns, of epic based deconvolution data}
\item{mcp}{matrix with 11 rows and 266 columns, of mcp-counter based deconvolution data}
\item{quantiseq}{matrix with 11 rows and 266 columns, of quantiseq based deconvolution data}

}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE119616}
}
\usage{
mae_kim
}
\description{
MultiAssayExperiment object containing gene expression (gex) and immune cell estimates from Kim et al.
}
\details{
The dataset consists of 266 NCCN very low/low or favorable-intermediate risk PCa patients
}
\references{
Kim HL, Li P, Huang HC, Deheshi S et al. Validation of the Decipher Test for predicting adverse pathology in candidates for prostate cancer active surveillance. Prostate Cancer Prostatic Dis 2019 Sep;22(3):399-405.
(\href{https://pubmed.ncbi.nlm.nih.gov/30542054/}{PubMed})
}
\keyword{datasets}
