% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_kunderfranco}
\alias{mae_kunderfranco}
\title{Kunderfranco et al. MAE-object}
\format{
An MAE object spanning 67 samples of normal prostate samples and prostate cancer samples
\describe{
  \item{gex.logr}{matrix with 16546 rows and 67 columns, of gene expression data}
  \item{cibersort}{matrix with 22 rows and 67 columns, of cibersort based deconvolution data}
  \item{xcell}{matrix with 39 rows and 67 columns, of xcell based deconvolution data}
  \item{epic}{matrix with 8 rows and 67 columns, of epic based deconvolution data}
  \item{quantiseq}{matrix with 11 rows and 67 columns, of quantiseq based deconvolution data}
  \item{mcp}{matrix with 11 rows and 67 columns, of mcp-counter based deconvolution data}
  \item{scores}{matrix with 2 rows and 67 columns, of Prolaris risk scores and AR scores}
  \item{purity}{matrix with 1 rows and 53 columns, of tumor purity estimates}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE14206}
}
\usage{
mae_kunderfranco
}
\description{
MultiAssayExperiment object containing gene expression (gex) and immune cell estimates from Kunderfranco et al.
}
\details{
The data contains 14 disease free benign prostate hyperplasia smaples and 53 prostate cancer samples.
}
\references{
Kunderfranco, P., Mello-Grand, M., Cangemi, R., et al.  (2010). ETS transcription factors control transcription of EZH2 and epigenetic silencing of the tumor suppressor gene Nkx3.1 in prostate cancer. PloS one, 5(5), e10547. https://doi.org/10.1371/journal.pone.0010547
 (\href{https://pubmed.ncbi.nlm.nih.gov/20479932/}{PubMed})
}
\keyword{datasets}
