% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_ren}
\alias{mae_ren}
\title{Ren 2017 MAE-object}
\format{
An MAE object spanning 112 men with prostate cancer
\describe{
  \item{gex}{matrix with 21589 rows and 65 columns}
  \item{cna}{matrix with 20873 rows and 65 columns}
}
}
\source{
\url{https://www.cbioportal.org/study/summary?id=prad_eururol_2017}
}
\usage{
mae_ren
}
\description{
MultiAssayExperiment object containing gene expression (gex) and copy 
number alteration (cna).
}
\keyword{datasets}
