% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_taylor}
\alias{mae_taylor}
\title{MAE object for Taylor et al.

TODO}
\format{
An object of class \code{MultiAssayExperiment} of length 4.
}
\usage{
mae_taylor
}
\description{
MAE object for Taylor et al.

TODO
}
\keyword{datasets}
