% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_tcga}
\alias{mae_tcga}
\title{MAE object for TCGA

TODO}
\format{
An object of class \code{MultiAssayExperiment} of length 2.
}
\usage{
mae_tcga
}
\description{
MAE object for TCGA

TODO
}
\keyword{datasets}
