% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_wallace}
\alias{mae_wallace}
\title{Wallace et al. MAE-object}
\format{
An MAE object spanning 333 men with prostate cancer
\describe{
  \item{gex}{matrix with 12048 rows and 83 columns}
  \item{mcp}{matrix with 11 rows and 83 columns}
  \item{quantiseq}{matrix with 4 rows and 83 columns}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE6956}
}
\usage{
mae_wallace
}
\description{
Wallace et al. MAE-object
}
\details{
The data contains 69 samples. For all 69 sample the dataset contains tumor data, and for 14 individual there is a matched healthy prostate tissue sample (the clinical data
}
\references{
Wallace, T. A., Prueitt, R. L., Yi, M., Howe, T. M., Gillespie, J. W., Yfantis, H. G., Stephens, R. M., Caporaso, N. E., Loffredo, C. A., & Ambs, S. (2008). Tumor immunobiological differences in prostate cancer between African-American and European-American men. Cancer research, 68(3), 927–936. https://doi.org/10.1158/0008-5472.CAN-07-2608
 (\href{https://pubmed.ncbi.nlm.nih.gov/18245496/}{PubMed})
}
\keyword{datasets}
