% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_wang}
\alias{mae_wang}
\title{Wang et al. MAE-object}
\format{
An MAE object spanning 148 men with prostate cancer
\describe{
  \item{gex.rma}{matrix with 12798 rows and 148 columns, of gene expression data}
  \item{cibersort}{matrix with 22 rows and 148 columns, of cibersort based deconvolution data}
  \item{xcell}{matrix with 39 rows and 148 columns, of xcell based deconvolution data}
  \item{epic}{matrix with 8 rows and 148 columns, of epic based deconvolution data}
  \item{quantiseq}{matrix with 11 rows and 148 columns, of quantiseq based deconvolution data}
  \item{mcp}{matrix with 11 rows and 148 columns, of mcp-counter based deconvolution data}
  \item{scores}{matrix with 4 rows and 148 columns, of risk scores and AR scores}
  \item{estimate}{data.frame with 4 rows and 148 columns, of cell types based on ESTIMATE method}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE8218}
}
\usage{
mae_wang
}
\description{
MultiAssayExperiment object containing gene expression (gex) and immune cell estimates from Wang et al.
}
\details{
148 prostate samples, with various amounts of tumor, stroma, BPH and atrophic gland, were used for this study.
}
\examples{
mae_wang
}
\references{
Wang Y, Xia XQ, Jia Z, Sawyers A et al. In silico estimates of tissue components in surgical samples based on expression profiling data. Cancer Res 2010 Aug 15;70(16):6448-55.
 (\href{https://pubmed.ncbi.nlm.nih.gov/20663908/}{PubMed})
}
\keyword{datasets}
