% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\docType{data}
\name{shiny}
\alias{shiny}
\alias{ui}
\alias{server}
\title{Launch R Shiny interface for the curatedPCaData-package}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.
}
\usage{
shiny()

ui

server(input, output, session)
}
\description{
Description

Description

Description
}
\keyword{datasets}
