% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shinyPCa}
\alias{shinyPCa}
\title{Launch R Shiny interface for the curatedPCaData-package}
\usage{
shinyPCa()
}
\value{
An interactive R Shiny browser session
}
\description{
shinyPCa launches an R Shiny web interface session, which can be used to access the mae-objects and data there-in.
}
\examples{
if(interactive()){
    shinyPCa()
}

}
