% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sun_gex}
\alias{sun_gex}
\title{Sun et al gene expression.}
\format{
A matrix with 12057 rows and 79 samples:
\describe{
  \item{samples}{raw gene expression values for that sample}
  ...
}
}
\usage{
sun_gex
}
\description{
A matrix containing the gene expression values  \url{https://www.ncbi.nlm.nih.gov/pubmed/19343730}
}
\keyword{datasets}
