% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taylor_cna}
\alias{taylor_cna}
\title{Taylor et al copy number alteration}
\format{
A matrix with 182 rows and 36630 genes:
\describe{
  \item{genes}{raw CNAa values}
  ...
}
}
\usage{
taylor_cna
}
\description{
A matrix containing the CNA values for the 182 samples from 
the Cancer Cell paper \url{https://pubmed.ncbi.nlm.nih.gov/20579941/}
}
\keyword{datasets}
