% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taylor_gex}
\alias{taylor_gex}
\title{Taylor et al gene expression.}
\format{
A matrix with 182 rows and 36630 genes:
\describe{
  \item{genes}{raw gene expression values}
  ...
}
}
\usage{
taylor_gex
}
\description{
A matrix containing the gene expression values for the 182 samples from 
the Cancer Cell paper \url{https://pubmed.ncbi.nlm.nih.gov/20579941/}
}
\keyword{datasets}
