% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_cna}
\alias{tcga_cna}
\title{TCGA copy number.}
\format{
A matrix with 333 rows and 36630 genes:
\describe{
  \item{genes}{copy number values}
  ...
}
}
\usage{
tcga_cna
}
\description{
A matrix containing the copy number values for the 333 samples included 
in the 2015 Cell paper \url{https://www.ncbi.nlm.nih.gov/pubmed/26544944}
}
\keyword{datasets}
