% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_gex}
\alias{tcga_gex}
\title{TCGA gene expression.}
\format{
A matrix with 333 rows and 36630 genes:
\describe{
  \item{genes}{raw gene expression values}
  ...
}
}
\usage{
tcga_gex
}
\description{
A matrix containing the gene expression values for the 333 samples included 
in the 2015 Cell paper \url{https://www.ncbi.nlm.nih.gov/pubmed/26544944}
}
\keyword{datasets}
