% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_prad}
\alias{template_prad}
\title{Template for curatedPCaData clinical fields}
\format{
An object of class \code{data.frame} with 68 rows and 6 columns.
}
\usage{
template_prad
}
\description{
This data frame contains the clinical fields which were the aim for extraction when gathering metadata for each dataset. 
Variable types, ranges, and description is reported.
}
\examples{
data(template_prad)
head(template_prad)
}
\keyword{datasets}
