% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateAnno}
\alias{updateAnno}
\title{Update gene annotations and resolve ambiguity}
\usage{
updateAnno(
  x,
  main = "hgnc_symbol",
  type = c("Aliases", "ensembl_gene_id", "ensembl_transcript_id", "refseq_mrna"),
  collapse_fun,
  omitNAempty = TRUE,
  ...
)
}
\description{
Update gene annotations and resolve ambiguity
}
