% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_mae}
\alias{create_mae}
\title{Create multi-assay experiment object.}
\usage{
create_mae(study_name = c("TCGA", "Taylor", "Sun"), verb = TRUE, ...)
}
\arguments{
\item{study_name}{study identifier.}

\item{verb}{print output as MAE is being created}

\item{...}{additional arguments}
}
\value{
A MultiAssayExperiment object containing clinical and multi-'omics data
}
\description{
Create multi-assay experiment object.
}
