% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_mae}
\alias{create_mae}
\title{Create multi-assay experiment object.}
\usage{
create_mae(study_name = "TCGA", verb = 1, ...)
}
\arguments{
\item{study_name}{study identifier.}
}
\value{
A MultiAssayExperiment object containing clinical and multi-'omics data
}
\description{
Create multi-assay experiment object.
}
