% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedPCaData-package.R
\docType{package}
\name{curatedPCaData-package}
\alias{curatedPCaData}
\alias{curatedPCaData-package}
\title{curatedPCaData: Curated Prostate Cancer Data}
\description{
A selection of annotated prostate cancer datasets including TCGA, 
    Taylor et al., Sun et al., and others.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Syksy/curatedPCaData}
  \item Report bugs at \url{https://github.com/Syksy/curatedPCaData/issues}
}

}
\author{
\strong{Maintainer}: Teemu Daniel Laajala \email{teelaa@utu.fi.com}

Other contributors:
\itemize{
  \item Jordan Creed \email{jordan.h.creed@moffitt.org} [contributor]
  \item Christelle Colin Leitzinger \email{christelle.colinleitzinger@moffitt.org} [contributor]
  \item Varsha Sreekanth \email{varsha.sreekanth@cuanschutz.edu} [contributor]
}

}
\keyword{internal}
