% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedPCaData-package.R
\docType{package}
\name{curatedPCaData}
\alias{curatedPCaData}
\alias{curatedPCaData-package}
\title{curatedPCaData}
\description{
The package curatedPCaData offers a selection of annotated
prostate cancer datasets featuring multiple omics, manually
curated metadata, and derived downstream variables. The studies
are offered as MultiAssayExperiment (MAE) objects via ExperimentHub,
and comprise of clinical characteristics tied to gene expression,
copy number alteration and somatic mutation data. Further, downstream
features computed from these multi-omics data are offered. Multiple
vignettes help grasp characteristics of the various studies and provide
example exploratory and meta-analysis of leveraging the multiple
studies provided here-in.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Syksy/curatedPCaData}
  \item Report bugs at \url{https://github.com/Syksy/curatedPCaData/issues}
}

}
\author{
\strong{Maintainer}: Teemu Daniel Laajala \email{teelaa@utu.fi} (\href{https://orcid.org/0000-0002-7016-7354}{ORCID})

Other contributors:
\itemize{
  \item Jordan Creed \email{jordan.h.creed@moffitt.org} [contributor]
  \item Christelle Colin Leitzinger \email{christelle.colinleitzinger@moffitt.org} [contributor]
  \item Varsha Sreekanth \email{varsha.sreekanth@cuanschutz.edu} [contributor]
  \item Federico Calboli \email{federico.calboli@luke.fi} [contributor]
  \item Kalaimathy Singaravelu \email{kalsin@utu.fi} [contributor]
  \item Michael Orman \email{michael.orman@cuanschutz.edu} [contributor]
  \item Alex Soupir \email{Alex.Soupir@moffitt.org} [contributor]
  \item Anni Halkola \email{ansuha@utu.fi} [contributor]
}

}
\keyword{internal}
