% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpca.R
\name{getPCa}
\alias{getPCa}
\title{Construct Prostate Cancer MultiAssayExperiment object from specific cohort}
\usage{
getPCa(dataset, slots, timestamp, verbose = FALSE, ...)
}
\arguments{
\item{dataset}{character() of PCa cancer cohort names
(e.g., 'abida')}

\item{slots}{character() A vector of PCa assays. If not included, returns all
available for the selected dataset;
see below for more details}

\item{timestamp}{character(1) "20230215" indicating the data version to obtain
from `ExperimentHub`. See `version` section details.}

\item{verbose}{logical(1) Whether to show the dataset currenlty being
(down)loaded (default TRUE)}

\item{...}{Additional arguments passed on to the
\code{\link[ExperimentHub:ExperimentHub-class]{ExperimentHub}}
constructor}
}
\value{
a \linkS4class{MultiAssayExperiment} of the specified assays and
cancer code
}
\description{
curatedPCaData provides \linkS4class{MultiAssayExperiment}
container objects that are constructed from ExperimentHub.
User provides PCa dataset name (see list on the overview-vignette or help files).
}
\details{
This function will check against available resources in
ExperimentHub.
For a list of datasets, see the overview-vignette.
}
\section{Available Assays}{


Thelist of ExperimentList assay names and their descriptions.
These assays can be entered as part of the \code{slots} argument in the
main function.
\preformatted{

ExperimentList data types   Description
----------------------------------------------------------------------------

  gex.rma            Gene expression values
  gex.logq
  gex.relz
  gex.logr
  gex.rsem.log
  cna.gistic         Copy number alteration
  cna.logr
  mut                Somatic mutations
  cibersort
  xcell
  epic
  quantiseq
  mcp
  estimate
  scores
}
}

\section{timestamp}{
 "20230215"
The timestamp is updated in case the data is updated. In this case, this
section describes the changes made for new timestamps. At this time, the
only data deposit is from 2023, Feb 15th, indicated by "20230215".
}

\examples{
mae_taylor <- getPCa(
  dataset = "taylor", timestamp = "20230215"
)

}
\seealso{
curatedPCaData-package
}
