% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_abida}
\alias{mae_abida}
\title{Abida et al. MAE-object}
\format{
A MAE object spanning 444 castrate resistant prostate cancer samples.
\describe{
   \item{cna.gistic}{a matrix with 20264 rows and 444 columns, from GISTIC discretized copy number alteration calls.}
   \item{gex.relz}{a matrix with 18823 rows and 266 columns, from z-score normalized expression in relative to paired normals.}
   \item{mut}{RaggedExperiment with 64566 rows and 444 columns, mutation data from cbioportal}
   \item{cibersort}{matrix with 22 rows and 266 columns, of cibersort based deconvolution data}
   \item{xcell}{matrix with 39 rows and 266 columns, of xcell based deconvolution data}
   \item{epic}{matrix with 8 rows and 266 columns, of epic based deconvolution data}
   \item{quantiseq}{matrix with 11 rows and 266 columns, of quantiseq based deconvolution data}
   \item{mcp}{matrix with 11 rows and 266 columns, of mcp-counter based deconvolution data}
   \item{scores}{matrix with 2 rows and 266 columns, of Prolaris risk scores and AR scores}
   
}
}
\source{
\url{https://www.cbioportal.org/study/summary?id=prad_su2c_2019}
}
\usage{
mae_abida
}
\description{
MultiAssayExperiment object containing gene expression (gex), copy 
number alteration (cna), mutations (mut) and immune cell estimates for Abida et al.
}
\details{
the clinical data refers to a sample of 444 tumors collected in 429 patients. The tissue was collected primarily at metastatic sites rather than from the prostate.
}
\references{
Abida, W., Cyrta, J., Heller, G., et al. (2019). Genomic correlates of clinical outcome in advanced prostate cancer. Proceedings of the National Academy of Sciences of the United States of America, 116, 11428 - 11436.
(\href{https://pubmed.ncbi.nlm.nih.gov/31061129/}{PubMed})
}
\keyword{datasets}
