% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_baca}
\alias{mae_baca}
\title{Baca et al.}
\format{
An object of class \code{MultiAssayExperiment} of length 1.
}
\source{
\url{...}
}
\usage{
mae_baca
}
\description{
MAE ...
}
\keyword{datasets}
