% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_baca}
\alias{mae_baca}
\title{Baca et al. MAE-object}
\format{
A MAE object spanning 56 prostate cancer samples.
\describe{
  \item{cna.gistic}{matrix with 19661 rows and 56 columns, from GISTIC discretized copy number alteration calls.}
}
}
\source{
\url{https://www.cbioportal.org/study/summary?id=prad_broad_2013}
}
\usage{
mae_baca
}
\description{
MultiAssayExperiment object containing copy number alteration (cna) for Baca et al.
}
\keyword{datasets}
