% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_barbieri}
\alias{mae_barbieri}
\title{Barbieri et al. MAE-object}
\format{
A MAE object spanning prostate cancer samples.
\describe{
  \item{cna.gistic}{matrix with 21723 rows and 109 columns, from GISTIC discretized copy number alteration calls.}
  \item{gex.relz}{matrix with 18193 rows and 20 columns, from z-score normalized expression in relative to paired normals.}
  \item{mut}{matrix with 29009 rows and 112 columns, genetic aberration calls.}
}
}
\source{
\url{https://www.cbioportal.org/study/summary?id=prad_broad}
}
\usage{
mae_barbieri
}
\description{
MultiAssayExperiment object for Barbieri et al.
}
\details{
The data comprises of primary localised prostate tumors from two cohorts, the Weill Cornell Medical College (WCMC; New York, NY), and the Uropath (Perth, Australia), which commercially provides banked urological tissues. None of the samples comes from patients who had received prior treatment for prostate cancer.
}
\references{
Barbieri, C. E., Baca, S. C., Lawrence, M. S., et al. (2012). Exome sequencing identifies recurrent SPOP, FOXA1 and MED12 mutations in prostate cancer. Nature genetics, 44(6), 685–689. https://doi.org/10.1038/ng.2279
(\href{https://pubmed.ncbi.nlm.nih.gov/22610119/}{PubMed})
}
\keyword{datasets}
