% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_barbieri}
\alias{mae_barbieri}
\title{Barbieri MAE-object}
\format{
An MAE object spanning 112 men with prostate cancer
\describe{
  \item{gex}{matrix with 18193 rows and 20 columns}
  \item{cna}{matrix with 21723 rows and 109 columns}
}
}
\source{
\url{https://www.cbioportal.org/study/summary?id=prad_broad}
}
\usage{
mae_barbieri
}
\description{
MultiAssayExperiment object containing gene expression (gex) and copy 
number alteration (cna).
}
\keyword{datasets}
