% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_chandran}
\alias{mae_chandran}
\title{Chandran et al., Yu et al. MAE-object}
\format{
An object of class \code{MultiAssayExperiment} of length 1.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE6919}
}
\usage{
mae_chandran
}
\description{
MultiAssayExperiment object containing GEX from Chandran et al., Yu et al. in GSE6919
}
\keyword{datasets}
