% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_chandran}
\alias{mae_chandran}
\title{Chandran et al., Yu et al. MAE-object}
\format{
An MAE object spanning 171 men with prostate cancer
\describe{
  \item{gex.rma}{matrix with 9007 rows and 171 columns, for the gene expression data}
  \item{cibersort}{matrix with 22 rows and 171 columns, of cibersort based deconvolution data}
  \item{xcell}{matrix with 39 rows and 171 columns, of xcell based deconvolution data}
  \item{epic}{matrix with 8 rows and 171 columns, of epic based deconvolution data}
  \item{quantiseq}{matrix with 11 rows and 171 columns, of quantiseq based deconvolution data}
  \item{mcp}{matrix with 11 rows and 171 columns, of mcp-counter based deconvolution data}
  \item{scores}{matrix with 2 rows and 171 columns, of Prolaris risk scores and AR scores}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE6919}
}
\usage{
mae_chandran
}
\description{
MultiAssayExperiment object containing gene expression (gex) and immune cell estimates for Chandran et al., Yu et al.
}
\details{
.
}
\references{
Chandran, U. R., Ma, C., Dhir, R., et al. (2007). Gene expression profiles of prostate cancer reveal involvement of multiple molecular pathways in the metastatic process. BMC cancer, 7, 64. https://doi.org/10.1186/1471-2407-7-64
(\href{https://pubmed.ncbi.nlm.nih.gov/17430594/}{PubMed})
Yu, Y. P., Landsittel, D., Jing, L., et al. (2004). Gene expression alterations in prostate cancer predicting tumor aggression and preceding development of malignancy. Journal of clinical oncology : official journal of the American Society of Clinical Oncology, 22(14), 2790–2799. https://doi.org/10.1200/JCO.2004.05.158
(\href{https://pubmed.ncbi.nlm.nih.gov/15254046/}{PubMed})
}
\keyword{datasets}
