% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_friedrich}
\alias{mae_friedrich}
\title{Friedrich et al. MAE-object}
\format{
An MAE object spanning 255 men with prostate cancer
\describe{
  \item{gex.logq}{matrix with 23097 rows and 255 columns, for the log-quantile normalized gene expression data}
  \item{cibersort}{matrix with 22 rows and 255 columns, of cibersort based deconvolution data}
  \item{xcell}{matrix with 39 rows and 255 columns, of xcell based deconvolution data}
  \item{epic}{matrix with 8 rows and 255 columns, of epic based deconvolution data}
  \item{quantiseq}{matrix with 11 rows and 255 columns, of quantiseq based deconvolution data}
  \item{mcp}{matrix with 11 rows and 255 columns, of mcp-counter based deconvolution data}
  \item{scores}{matrix with 2 rows and 255 columns, of Prolaris risk scores and AR scores}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE134051}
}
\usage{
mae_friedrich
}
\description{
MultiAssayExperiment object containing gene expression (gex) and immune cell estimates for Friedrich et al.
}
\details{
The data comprises of 255 samples, with 164 primary tumors samples, 52 adjacent normal samples, and 39 benign prostate hyperplasia samples.  This dataset includes in its totality the 164 samples analysed in Kreutz et al. 2020.
}
\references{
Friedrich, M., Wiedemann, K., Reiche, K., et al. (2020). The Role of lncRNAs TAPIR-1 and -2 as Diagnostic Markers and Potential Therapeutic Targets in Prostate Cancer. Cancers, 12(5), 1122. https://doi.org/10.3390/cancers12051122
(\href{https://pubmed.ncbi.nlm.nih.gov/32365858/}{PubMed})
Kreuz, M., Otto, D. J., Fuessel, et al. (2020). ProstaTrend-A Multivariable Prognostic RNA Expression Score for Aggressive Prostate Cancer. European urology, 78(3), 452–459. https://doi.org/10.1016/j.eururo.2020.06.001
(\href{https://pubmed.ncbi.nlm.nih.gov/32631745/}{PubMed})
}
\keyword{datasets}
