% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_hieronymus}
\alias{mae_hieronymus}
\title{Hieronymus et al. MAE-object}
\format{
A MAE object spanning 104 tumor samples
\describe{
  \item{cna.logr}{matrix with 22895 rows and 104 columns, for the log-ratios for copy number alteration data}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE54691}
}
\usage{
mae_hieronymus
}
\description{
MultiAssayExperiment object containing copy number alteration (cna) for Hieronymus et al.
}
\details{
The data comprises of 104 samples, for which are available clinical data and Copy Number Alteration, but no gene expression data -- thus no deconvolution results are available.
}
\examples{
mae_hieronymus
}
\references{
Hieronymus, H., Schultz, N., Gopalan, A., et al. (2014). Copy number alteration burden predicts prostate cancer relapse. Proceedings of the National Academy of Sciences of the United States of America, 111(30), 11139–11144. https://doi.org/10.1073/pnas.1411446111
(\href{https://pubmed.ncbi.nlm.nih.gov/25024180/}{PubMed})
}
\keyword{datasets}
