% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_ren}
\alias{mae_ren}
\title{Ren 2017 MAE-object}
\format{
An MAE object spanning 112 men with prostate cancer
\describe{
  \item{gex}{matrix with 21589 rows and 65 columns}
  \item{cna}{matrix with 20873 rows and 65 columns}
  \item{xcell}{matrix with 39 rows and 65 columns, of xcell based deconvolution data}
  \item{epic}{matrix with 8 rows and 65 columns, of epic based deconvolution data}
  \item{quantiseq}{matrix with 11 rows and 65 columns, of quantiseq based deconvolution data}
  \item{mcp}{matrix with 11 rows and 65 columns, of mcp-counter based deconvolution data}
}

An object of class \code{MultiAssayExperiment} of length 3.
}
\source{
\url{https://www.cbioportal.org/study/summary?id=prad_eururol_2017}

\url{https://www.cbioportal.org/study/summary?id=prad_eururol_2017}
}
\usage{
mae_ren

mae_ren
}
\description{
Ren 2017 MAE-object

Weiner 2021 MAE-object
}
\details{
The data contains 14 disease free benign prostate hyperplasia smaples and 53 prostate cancer samples.
}
\references{
Ren, S., Wei, G. H., Liu, D., et al.  (2018). Whole-genome and Transcriptome Sequencing of Prostate Cancer Identify New Genetic Alterations Driving Disease Progression. European urology, 73(3), 322–339. https://doi.org/10.1016/j.eururo.2017.08.027
 (\href{https://pubmed.ncbi.nlm.nih.gov/28927585/}{PubMed})
}
\keyword{datasets}
