% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_ren}
\alias{mae_ren}
\title{Ren et al. MAE-object}
\format{
An MAE object spanning 65 men with prostate cancer
\describe{
  \item{cna.gistic}{matrix with 20873 rows and 65 columns, from GISTIC discretized copy number alteration calls}
  \item{gex.relz}{matrix with 21589 rows and 65 columns, of gene expression data}
  \item{mut}{RaggedExperiment with 50803 rows and 65 columns, of mutation data from cbioportal}
  \item{cibersort}{matrix with 22 rows and 65 columns, of cibersort based deconvolution data}
  \item{xcell}{matrix with 39 rows and 65 columns, of xcell based deconvolution data}
  \item{epic}{matrix with 8 rows and 65 columns, of epic based deconvolution data}
  \item{quantiseq}{matrix with 11 rows and 65 columns, of quantiseq based deconvolution data}
  \item{mcp}{matrix with 11 rows and 65 columns, of mcp-counter based deconvolution data}
  \item{scores}{matrix with 2 rows and 65 columns, of Prolaris risk scores and AR scores}
}
}
\source{
\url{https://www.cbioportal.org/study/summary?id=prad_eururol_2017}
}
\usage{
mae_ren
}
\description{
MultiAssayExperiment object containing gene expression (gex), copy 
number alteration (cna), mutations and immune cell estimates from Ren et al.
}
\details{
The data contains 14 disease free benign prostate hyperplasia smaples and 53 prostate cancer samples.
}
\references{
Ren, S., Wei, G. H., Liu, D., et al.  (2018). Whole-genome and Transcriptome Sequencing of Prostate Cancer Identify New Genetic Alterations Driving Disease Progression. European urology, 73(3), 322–339. https://doi.org/10.1016/j.eururo.2017.08.027
 (\href{https://pubmed.ncbi.nlm.nih.gov/28927585/}{PubMed})
}
\keyword{datasets}
