% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_sun}
\alias{mae_sun}
\title{MAE object for Sun et al.

TODO}
\format{
An object of class \code{MultiAssayExperiment} of length 1.
}
\usage{
mae_sun
}
\description{
MAE object for Sun et al.

TODO
}
\keyword{datasets}
