% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_sun}
\alias{mae_sun}
\title{Sun et al. MAE-object}
\format{
An MAE object spanning 79 men with prostate cancer
\describe{
  \item{gex.rma}{matrix with 12798 rows and 79 columns, of gene expression data}
  \item{cibersort}{matrix with 22 rows and 79 columns, of cibersort based deconvolution data}
  \item{xcell}{matrix with 39 rows and 79 columns, of xcell based deconvolution data}
  \item{epic}{matrix with 8 rows and 79 columns, of epic based deconvolution data}
  \item{quantiseq}{matrix with 11 rows and 79 columns, of quantiseq based deconvolution data}
  \item{estimate}{data.frame with 4 rows and 79 columns, of cell types based on ESTIMATE method}
  \item{scores}{matrix with 4 rows and 79 columns, of risk scores and AR scores}
  \item{mcp}{matrix with 11 rows and 79 columns, of mcp-counter based deconvolution data}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE25136}
}
\usage{
mae_sun
}
\description{
MultiAssayExperiment object containing gene expression (gex) and immune cell estimates from Sun et al.
}
\examples{
mae_sun
}
\references{
Sun, Y., & Goodison, S. (2009). Optimizing molecular signatures for predicting prostate cancer recurrence. The Prostate, 69(10), 1119–1127. https://doi.org/10.1002/pros.20961
 (\href{https://pubmed.ncbi.nlm.nih.gov/19343730/}{PubMed})
}
\keyword{datasets}
