% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_taylor}
\alias{mae_taylor}
\title{Taylor et al. MAE-object}
\format{
An object of class \code{MultiAssayExperiment} of length 2.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE25136}
}
\usage{
mae_taylor
}
\description{
MultiAssayExperiment object containing GEX (exon and transcript) and CNA from Taylor et al.
}
\keyword{datasets}
