% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_tcga}
\alias{mae_tcga}
<<<<<<< HEAD
\title{TCGA MAE-object}
\format{
An MAE object spanning 333 men with prostate cancer
\describe{
  \item{GEX}{matrix with 19985 rows and 333 columns}
  \item{CNA}{matrix with 21761 rows and 333 columns}
}
}
\source{
\url{https://www.cbioportal.org/study/summary?id=prad_tcga_pub}
=======
\title{MAE object for TCGA

TODO}
\format{
An object of class \code{MultiAssayExperiment} of length 2.
>>>>>>> db064e9bd1d04bd273fcbab8e201097e6c6a8d3f
}
\usage{
mae_tcga
}
\description{
<<<<<<< HEAD
MultiAssayExperiment object containing GEX and CNA from TCGA
=======
MAE object for TCGA

TODO
>>>>>>> db064e9bd1d04bd273fcbab8e201097e6c6a8d3f
}
\keyword{datasets}
