% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_tcga}
\alias{mae_tcga}
\title{TCGA MAE-object}
\format{
An MAE object spanning 333 men with prostate cancer
\describe{
  \item{gex}{matrix with 19958 rows and 333 columns}
  \item{cna}{matrix with 21735 rows and 333 columns}
  \item{xcell}{matrix with 39 rows and 333 columns}
  \item{epic}{matrix with 8 rows and 333 columns}
  \item{quantiseq}{matrix with 11 rows and 333 columns}
  \item{mcp}{matrix with 11 rows and 333 columns}
  \item{cibersortx}{matrix with 25 rows and 333 columns}
  \item{osfgex}{matrix with 58684 rows and 483 columns}
  \item{osf_xcell}{matrix with 39 rows and 483 columns}
  \item{osf_epic}{matrix with 8 rows and 483 columns}
  \item{osf_quantiseq}{matrix with 11 rows and 483 columns}
  \item{osf_mcp}{matrix with 11 rows and 483 columns}
  \item{osf_cibersortx}{matrix with 25 rows and 483 columns}
}
}
\source{
\url{https://www.cbioportal.org/study/summary?id=prad_tcga_pub}
}
\usage{
mae_tcga
}
\description{
TCGA MAE-object
}
\references{
Cancer Genome Atlas Research Network (2015). The Molecular Taxonomy of Primary Prostate Cancer. Cell, 163(4), 1011–1025. https://doi.org/10.1016/j.cell.2015.10.025
 (\href{https://pubmed.ncbi.nlm.nih.gov/26544944/}{PubMed})
}
\keyword{datasets}
