% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_tcga}
\alias{mae_tcga}
\title{TCGA MAE-object}
\format{
An MAE object spanning 333 men with prostate cancer
\describe{
  \item{gex}{matrix with 19985 rows and 333 columns}
  \item{cna}{matrix with 21761 rows and 333 columns}
}
}
\source{
\url{https://www.cbioportal.org/study/summary?id=prad_tcga_pub}
}
\usage{
mae_tcga
}
\description{
MultiAssayExperiment object containing GEX and CNA from TCGA
}
\keyword{datasets}
