% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_true}
\alias{mae_true}
\title{True et al. MAE-object}
\format{
An MAE object spanning 333 men with prostate cancer
\describe{
  \item{gex}{matrix with 3398 rows and 32 columns}
  \item{scores}{matrix with 2 rows and 32 columns, genomic risk and AR scores}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE5132}
}
\usage{
mae_true
}
\description{
MultiAssayExperiment object containing GEX from True et al.
}
\details{
The expression data for this dataset has been produced by a two colour chip that provides the relative expression levels for each tumor rather than its absolute value as in all other dataset.  Thus it is not possible to provide immune deconvolution for this dataset.  This dataset also has a rich clinical data.  In this dataset the Gleason grade group follows the grading described in \url{https://www.cancerresearchuk.org/about-cancer/prostate-cancer/stages/grades}. The clinical data also includes an extra column named 'gleason_group' where the the samples are grouped based on their primary and secondary greason scores as <=6|3+4|4+3|>=8.
}
\references{
True, L., Coleman, I., Hawley, S., et al. (2006). A molecular correlate to the Gleason grading system for prostate adenocarcinoma. Proceedings of the National Academy of Sciences of the United States of America, 103(29), 10991–10996. https://doi.org/10.1073/pnas.0603678103
 (\href{https://pubmed.ncbi.nlm.nih.gov/16829574/}{PubMed})
}
\keyword{datasets}
