% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_true}
\alias{mae_true}
\title{True et al. MAE-object}
\format{
An MAE object spanning 32 men with prostate cancer
\describe{
  \item{gex.logr}{matrix with 3615 rows and 32 columns, of gene expression data}
  \item{cibersort}{matrix with 22 rows and 32 columns, of cibersort based deconvolution data}
  \item{epic}{matrix with 8 rows and 32 columns, of epic based deconvolution data}
  \item{quantiseq}{matrix with 11 rows and 32 columns, of quantiseq based deconvolution data}
  \item{mcp}{matrix with 7 rows and 32 columns, of mcp-counter based deconvolution data}
  \item{scores}{matrix with 2 rows and 32 columns, of Prolaris risk scores and AR scores}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE5132}
}
\usage{
mae_true
}
\description{
MultiAssayExperiment object containing gene expression (gex) and immune cell estimates from True et al.
}
\details{
The expression data for this dataset was produced by an older two colour chip that provides the relative expression levels for between tumor and normal.  Due to the chip type, there was a lack of gene overlap with some of the downstream methods, thus some downstream methods are missing.
}
\examples{
mae_true
}
\references{
True, L., Coleman, I., Hawley, S., et al. (2006). A molecular correlate to the Gleason grading system for prostate adenocarcinoma. Proceedings of the National Academy of Sciences of the United States of America, 103(29), 10991–10996. https://doi.org/10.1073/pnas.0603678103
 (\href{https://pubmed.ncbi.nlm.nih.gov/16829574/}{PubMed})
}
\keyword{datasets}
