% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_wallace}
\alias{mae_wallace}
\title{Wallace et al. MAE-object}
\format{
An object of class \code{MultiAssayExperiment} of length 1.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE6956}
}
\usage{
mae_wallace
}
\description{
MultiAssayExperiment object containing GEX from Wallace et al.
}
\keyword{datasets}
