% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_wallace}
\alias{mae_wallace}
\title{Wallace et al. MAE-object}
\format{
An MAE object spanning 83 men with prostate cancer
\describe{
  \item{gex.rma}{matrix with 12783 rows and 89 columns, of gene expression data}
  \item{cibersort}{matrix with 22 rows and 69 columns, of cibersort based deconvolution data}
  \item{xcell}{matrix with 39 rows and 75 columns, of xcell based deconvolution data}
  \item{epic}{matrix with 8 rows and 75 columns, of epic based deconvolution data}
  \item{quantiseq}{matrix with 11 rows and 75 columns, of quantiseq based deconvolution data}
  \item{mcp}{matrix with 11 rows and 75 columns, of mcp-counter based deconvolution data}
  \item{scores}{matrix with 4 rows and 75 columns, of risk scores and AR scores}
  \item{estimate}{data.frame with 4 rows and 75 columns, of cell types based on ESTIMATE method}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE6956}
}
\usage{
mae_wallace
}
\description{
MultiAssayExperiment object containing gene expression (gex) and immune cell estimates from Wallace et al.
}
\examples{
mae_wallace
}
\references{
Wallace, T. A., Prueitt, R. L., Yi, M., Howe, T. M., Gillespie, J. W., Yfantis, H. G., Stephens, R. M., Caporaso, N. E., Loffredo, C. A., & Ambs, S. (2008). Tumor immunobiological differences in prostate cancer between African-American and European-American men. Cancer research, 68(3), 927–936. https://doi.org/10.1158/0008-5472.CAN-07-2608
 (\href{https://pubmed.ncbi.nlm.nih.gov/18245496/}{PubMed})
}
\keyword{datasets}
