% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_weiner}
\alias{mae_weiner}
\title{Weiner et al. MAE-object}
\format{
An MAE spanning 838 prostate cancer samples of two cohorts
\describe{
 \item{gex.rma}{matrix of 17410 rows and 838 columns of gene expression data}
 \item{mcp}{matrix with 11 rows and 838 columns, the mcp-counter deconvolution of the expression data}
 \item{quantiseq}{matrix with 11 rows and 838 columns, the quantiseq deconvolution of the expression data}
 \item{xcell}{matrix with 39 rows and 838 columns, the xcell deconvolution of the expression data}
 \item{epic}{matrix with 8 rows and 838 columns, the epic deconvolution of the expression data}
 \item{scores}{matrix with 2 rows and 838 columns, of Prolaris risk scores and AR scores}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE157548}
}
\usage{
mae_weiner
}
\description{
A MultiAssayExperiment object containing gene expression (gex) and immune cell estimates for Weiner et al.
}
\references{
Weiner, A.B., Vidotto, T., Liu, Y. et al. Plasma cells are enriched in localized prostate cancer in Black men and are associated with improved outcomes. Nat Commun 12, 935 (2021). https://doi.org/10.1038/s41467-021-21245-w
(\href{https://pubmed.ncbi.nlm.nih.gov/33568675/}{PubMed})
}
\keyword{datasets}
