% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mae_weiner}
\alias{mae_weiner}
\title{Weiner et al. MAE-object}
\format{

}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE157548}
}
\usage{
mae_weiner
}
\description{
A MultiAssayExperiment object containing gene expression (gex) and immune cell estimates for Weiner et al.
}
\references{
Weiner, A.B., Vidotto, T., Liu, Y. et al. Plasma cells are enriched in localized prostate cancer in Black men and are associated with improved outcomes. Nat Commun 12, 935 (2021). https://doi.org/10.1038/s41467-021-21245-w
(\href{https://pubmed.ncbi.nlm.nih.gov/33568675/}{PubMed})
}
\keyword{datasets}
